List known updates on startup
Summary:
Currently it's possible that it just says "No updates available" although
PackageKit knows the opposite. This is because it doesn't even ask PK on
startup. With this, it always lists available updates, only refreshing
the cache if necessary.

Test Plan:
Ran plasmoidviewer, now lists available updates instead of
"Your system is up to date".

Reviewers: lukas, jgrulich

Subscribers: plasma-devel

Differential Revision: https://phabricator.kde.org/D9508