Fix expansion handling of PackageDelegate items
Kirigami's AbstractListItem handles setting the item as current, which
conflicted with the onClick handler in Full.qml. Adjust it so that it can
deal with that.

Additionally, PackageDelegate is a Button, so the checked state was overwritten
on clicks, which made the highlight disappear when selecting. The "highlighted"
property seems to not work here, so add a workaround.