Added a package deselection config option
It may seem redundant but it may be needed in some distros.
For instance, in openSUSE 15.3, PackageKit does not yet support libzypp locks.
It will allow updates to be scheduled without deselecting packages manually.
The deselected packages will still appear in the list and count towards the
updates pending, however.